pyproject.toml
[project]
name = "20-agi-copywriter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.0.0",
    "python-dotenv>=1.1.1",
    "groq>=0.4.0",
    "openai>=1.99.2",
    "tavily-python>=0.7.10",
    "sqlalchemy>=2.0.42",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "ipykernel>=6.30.1",
]
